requirements.txt
mxnet-cu92 >= 1.3.0
gluonnlp
sklearn
spacy